+++ naga/src/front/wgsl/mod.rs
@@ -11,7 +11,6 @@ mod parse;
 #[cfg(test)]
 mod tests;
 
-pub use crate::front::wgsl::error::ParseError;
 pub use crate::front::wgsl::parse::directive::language_extension::{
     ImplementedLanguageExtension, LanguageExtension, UnimplementedLanguageExtension,
 };
@@ -20,9 +19,11 @@ use alloc::boxed::Box;
 use thiserror::Error;
 
 use crate::front::wgsl::error::Error;
+pub use crate::front::wgsl::error::ParseError;
 use crate::front::wgsl::lower::Lowerer;
 use crate::front::wgsl::parse::Parser;
 use crate::Scalar;
+pub use parse::Options;
 
 #[cfg(test)]
 use std::println;
@@ -31,12 +32,20 @@ pub(crate) type Result<'a, T> = core::result::Result<T, Box<Error<'a>>>;
 
 pub struct Frontend {
     parser: Parser,
+    options: Options,
 }
 
 impl Frontend {
     pub const fn new() -> Self {
         Self {
             parser: Parser::new(),
+            options: Options::new(),
+        }
+    }
+    pub const fn new_with_options(options: Options) -> Self {
+        Self {
+            parser: Parser::new(),
+            options,
         }
     }
 
@@ -45,7 +54,7 @@ impl Frontend {
     }
 
     fn inner<'a>(&mut self, source: &'a str) -> Result<'a, crate::Module> {
-        let tu = self.parser.parse(source)?;
+        let tu = self.parser.parse(source, &self.options)?;
         let index = index::Index::generate(&tu)?;
         let module = Lowerer::new(&index).lower(tu)?;